requirements.txt
# 核心依赖
openai>=1.0.0
python-dotenv>=1.0.0

# PDF 解析
PyMuPDF>=1.23.0

# 向量数据库
faiss-cpu>=1.7.4
numpy>=1.24.0

# 工具依赖
arxiv>=2.0.0
duckduckgo-search>=3.9.0
pyyaml>=6.0.0

# 终端渲染
rich>=13.0.0
tqdm>=4.65.0

# MCP 协议支持（可选）
mcp>=1.0.0; extra == "mcp"

# 类型提示
pydantic>=2.0.0
